+++ src/Scene.js
@@ -74,8 +74,6 @@ export default class Scene extends Phaser.Scene {
 		this[HOOKS] = hooks;
 		this[PROPS] = props;
 
-		window.sc = this;
-
 		this.update = hooks.update || this.update;
 	}
 
@@ -123,7 +121,9 @@ export default class Scene extends Phaser.Scene {
 		beforecreate && beforecreate.call(this, this, ...args);
 
 		for (let i = 0, l = pool.length; i < l; i++) {
-			pool[i].register(this);
+			if (!pool[i].destroyed) {
+				pool[i].register(this);
+			}
 		}
 
 		create && create.call(this, this, ...args);