+++ src/objects/Audio.js
@@ -20,9 +20,12 @@ class Audio extends GameObject {
 	register(scene) {
 		const { name } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = scene.sound.add(name);
 		this.registered = true;
 		this.update(this.props);
+		this.postRegister();
+
 		return this.instance;
 	}
 }