+++ src/objects/Sprite.js
@@ -46,20 +46,24 @@ const performedProps = {
 	origin
 };
 
+const eventMap = {
+	onAnimationComplete: 'animationcomplete'
+};
+
 class Sprite extends GameObject {
 	register(scene) {
 		const { x, y, texture, frame } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = new Phaser.GameObjects.Sprite(scene, x, y, texture, frame);
 		this.registered = true;
 		scene.add.displayList.add(this.instance);
 		scene.add.updateList.add(this.instance);
+		this.texture = texture;
 		this.registerTransitions();
 		this.registerAnimations();
 		this.update(this.props);
-
-		window.sprite = this;
-
+		this.postRegister();
 		return this.instance;
 	}
 
@@ -120,7 +124,8 @@ Object.assign(Sprite.prototype, {
 	type: TYPES.SPRITE,
 	allowedProps,
 	performedProps,
-	transitionProps
+	transitionProps,
+	eventMap
 });
 
 export default Sprite;