+++ src/objects/Triangle.js
@@ -37,10 +37,11 @@ class Triangle extends GameObject {
 	register(scene) {
 		const { x, y, x1, y1, x2, y2, x3, y3, fillColor } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = scene.add.triangle(x, y, x1, y1, x2, y2, x3, y3, fillColor);
 		this.registered = true;
-		window.triangle = this;
 		this.update(this.props);
+		this.postRegister();
 
 		return this.instance;
 	}