+++ src/objects/Tween.js
@@ -23,8 +23,6 @@ class Tween extends TransparentGameObject {
 		this.tweenQueue = [];
 		this.animationPool = {};
 
-		window.tween = this;
-
 		this.animationsConfig = this.prepareConfig(animations);
 	}
 
@@ -44,7 +42,6 @@ class Tween extends TransparentGameObject {
 
 	play(key) {
 		const { replaceAnimation } = this.props;
-		console.log(key, this.animationsConfig);
 		const { queue, complex } = this.animationsConfig[key];
 
 		if (complex) {