+++ CHANGELOG.md
@@ -4,9 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://gitlab.com/firanolfind/react-phaser-bindings/compare/develop...master) (2019-12-03)
+## [Unreleased](https://gitlab.com/firanolfind/react-phaser-bindings/compare/develop...master) (2019-12-10)
 
 _put moar here_
+
+## [v1.3.0](https://gitlab.com/firanolfind/react-phaser-bindings/compare/v1.2.1...v1.3.0) (2019-12-10)
+
+### Added
+
+- **CounterText** component
+- **CounterBitmapText** component
+
+### Changed
+
+- [internal] added #preRegister method to add game objects
+- [internal] removed window debug variables
 
 ## [v1.2.1](https://gitlab.com/firanolfind/react-phaser-bindings/compare/v1.2.0...v1.2.1) (2019-12-03)